Exclude text widget editor
If you enter two or more paragraphs in visual mode, it adds a linebreak after the closing </p> tags, instead of writing everything in one line. This repeats when switching between the visual and text tab, and the so generated empty lines get converted into <p>&nbsp;</p>. As the text widget shouldn't be used for complex Markup, it's better to remove the plugins functionality here (on the customizer screen it doesn't apply anyway).